requirements.txt
Pillow
keyboard
tkinterdnd2
python-vlc